Address review: move validator registry to relations/, rename placement field
- Move PlacementValidatorRegistry and register_validator out of assets/ into
  relations/placement_validator_registry.py; they take no part in the asset
  cascade, so assets/registries.py and assets/register.py no longer reference them.
- Rename ArenaEnvGraphSpec.placement to placement_validators for clarity.
